% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASMK-package.R
\docType{package}
\name{ASMK-package}
\alias{ASMK}
\alias{ASMK-package}
\title{ASMK: Accelerated Spatial Meta-Kriging (ASMK)}
\description{
"This package provides the computational framework to fit the Accelerated Spatial Meta-Kriging (ASMK) approach presented and detailed in Bayesian Transfer Learning and Divide-Conquer Models for Massive Spatial Datasets of Presicce L., and Banerjee S. (2024) \doi{10.48550/arXiv.0000.0000}. The functions in the package use mostly Rcpp-armadillo implementations, offering computationally efficient Bayesian geostatistical modeling."
}
\author{
\strong{Maintainer}: Luca Presicce \email{l.presicce@campus.unimib.it} (\href{https://orcid.org/0009-0005-7062-3523}{ORCID})

}
\keyword{internal}
