% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BPS_latent_SpatialPrediction_cpp}
\alias{BPS_latent_SpatialPrediction_cpp}
\title{Compute the BPS spatial prediction given a set of stacking weights}
\usage{
BPS_latent_SpatialPrediction_cpp(
  data,
  X_u,
  priors,
  coords,
  crd_u,
  hyperpar,
  W,
  R
)
}
\arguments{
\item{data}{...}

\item{X_u}{...}

\item{priors}{...}

\item{coords}{...}

\item{crd_u}{...}

\item{hyperpar}{...}

\item{W}{...}

\item{R}{...}
}
\value{
list of BPS posterior predictive samples
}
\description{
Compute the BPS spatial prediction given a set of stacking weights
}
