% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BPS_post_draws}
\alias{BPS_post_draws}
\title{Compute the BPS posterior samples given a set of stacking weights}
\usage{
BPS_post_draws(data, priors, coords, hyperpar, W, R)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{priors}{\link{list} priors: named \eqn{\mu_b},\eqn{V_b},\eqn{a},\eqn{b}}

\item{coords}{\link{matrix} sample coordinates for X and Y}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\delta}, second named \eqn{\phi}}

\item{W}{\link{matrix} set of stacking weights}

\item{R}{\link{integer} number of desired samples}
}
\value{
\link{matrix} BPS posterior samples
}
\description{
Compute the BPS posterior samples given a set of stacking weights
}
