% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BPS_post_draws_latent}
\alias{BPS_post_draws_latent}
\title{Compute the BPS posterior samples given a set of stacking weights}
\usage{
BPS_post_draws_latent(data, priors, coords, hyperpar, W, R)
}
\arguments{
\item{data}{\link{list} two elements: first named\eqn{Y}, second named\eqn{X}}

\item{priors}{\link{list} priors: named\eqn{\mu_B},\eqn{V_r},\eqn{\Psi},\eqn{\nu}}

\item{coords}{\link{matrix} sample coordinates for X and Y}

\item{hyperpar}{\link{list} two elemets: first named\eqn{\alpha}, second named\eqn{\phi}}

\item{W}{\link{matrix} set of stacking weights}

\item{R}{\link{integer} number of desired samples}
}
\value{
\link{list} BPS posterior samples
}
\description{
Compute the BPS posterior samples given a set of stacking weights
}
