% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BPS_postdraws_MvT}
\alias{BPS_postdraws_MvT}
\title{Compute the BPS posterior samples given a set of stacking weights}
\usage{
BPS_postdraws_MvT(data, priors, coords, hyperpar, W, R, par)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{priors}{\link{list} priors: named \eqn{\mu_B},\eqn{V_r},\eqn{\Psi},\eqn{\nu}}

\item{coords}{\link{matrix} sample coordinates for X and Y}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\alpha}, second named \eqn{\phi}}

\item{W}{\link{matrix} set of stacking weights}

\item{R}{\link{integer} number of desired samples}

\item{par}{if \code{TRUE} only \eqn{\beta} and \eqn{\Sigma} are sampled (\eqn{\omega} is omitted)}
}
\value{
\link{matrix} BPS posterior samples
}
\description{
Compute the BPS posterior samples given a set of stacking weights
}
