% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BPS_PseudoBMA}
\alias{BPS_PseudoBMA}
\title{Combine subset models wiht Pseudo-BMA}
\usage{
BPS_PseudoBMA(fit_list)
}
\arguments{
\item{fit_list}{\link{list} K fitted model outputs composed by two elements each: first named \eqn{epd}, second named \eqn{W}}
}
\value{
\link{matrix} posterior predictive density evaluations (each columns represent a different model)
}
\description{
Combine subset models wiht Pseudo-BMA
}
