% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BPS_weights}
\alias{BPS_weights}
\title{Compute the BPS weights by convex optimization}
\usage{
BPS_weights(data, priors, coords, hyperpar, K)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{priors}{\link{list} priors: named \eqn{\mu_b},\eqn{V_b},\eqn{a},\eqn{b}}

\item{coords}{\link{matrix} sample coordinates for X and Y}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\delta}, second named \eqn{\phi}}

\item{K}{\link{integer} number of folds}
}
\value{
\link{matrix} posterior predictive density evaluations (each columns represent a different model)
}
\description{
Compute the BPS weights by convex optimization
}
