% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVXR_opt}
\alias{CVXR_opt}
\title{Compute the BPS weights by convex optimization}
\usage{
CVXR_opt(scores)
}
\arguments{
\item{scores}{\link{matrix} \eqn{N \times K} of expected predictive density evaluations for the K models considered}
}
\value{
conv_opt \link{function} to perform convex optimiazion with CVXR R package
}
\description{
Compute the BPS weights by convex optimization
}
