% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_pred_cpp}
\alias{d_pred_cpp}
\title{Evaluate the density of a set of unobserved response with respect to the conditional posterior predictive}
\usage{
d_pred_cpp(data, X_u, Y_u, Rphi_s, d_u, d_us, hyperpar, poster)
}
\arguments{
\item{data}{...}

\item{X_u}{...}

\item{Y_u}{...}

\item{Rphi_s}{...}

\item{d_u}{...}

\item{d_us}{...}

\item{hyperpar}{...}

\item{poster}{...}
}
\value{
vec of posterior predictive density evaluations
}
\description{
Evaluate the density of a set of unobserved response with respect to the conditional posterior predictive
}
