% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_pred_latent_cpp}
\alias{d_pred_latent_cpp}
\title{Evaluate the density of a set of unobserved response with respect to the conditional posterior predictive}
\usage{
d_pred_latent_cpp(data, X_u, Y_u, d_u, d_us, hyperpar, poster, beta, sigma)
}
\arguments{
\item{data}{...}

\item{X_u}{...}

\item{Y_u}{...}

\item{d_u}{...}

\item{d_us}{...}

\item{hyperpar}{...}

\item{poster}{...}

\item{beta}{...}

\item{sigma}{...}
}
\value{
posterior predictive density evaluation (double)
}
\description{
Evaluate the density of a set of unobserved response with respect to the conditional posterior predictive
}
