% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dens_kcv2}
\alias{dens_kcv2}
\title{Compute the KCV of the density evaluations for fixed values of the hyperparameters}
\usage{
dens_kcv2(data, priors, coords, hyperpar, K)
}
\arguments{
\item{data}{...}

\item{priors}{...}

\item{coords}{...}

\item{hyperpar}{...}

\item{K}{...}
}
\value{
list of posterior predictive density evaluations
}
\description{
Compute the KCV of the density evaluations for fixed values of the hyperparameters
}
