% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dens_kcv_MvT}
\alias{dens_kcv_MvT}
\title{Compute the KCV of the density evaluations for fixed values of the hyperparameters}
\usage{
dens_kcv_MvT(data, priors, coords, hyperpar, K)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{priors}{\link{list} priors: named \eqn{\mu_B},\eqn{V_r},\eqn{\Psi},\eqn{\nu}}

\item{coords}{\link{matrix} sample coordinates for X and Y}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\alpha}, second named \eqn{\phi}}

\item{K}{\link{integer} number of folds}
}
\value{
\link{vector} posterior predictive density evaluations
}
\description{
Compute the KCV of the density evaluations for fixed values of the hyperparameters
}
