% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dens_loocv_latent}
\alias{dens_loocv_latent}
\title{Compute the LOOCV of the density evaluations for fixed values of the hyperparameters}
\usage{
dens_loocv_latent(data, priors, coords, hyperpar)
}
\arguments{
\item{data}{...}

\item{priors}{...}

\item{coords}{...}

\item{hyperpar}{...}
}
\value{
vec of posterior predictive density evaluations
}
\description{
Compute the LOOCV of the density evaluations for fixed values of the hyperparameters
}
