% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expand_grid_cpp}
\alias{expand_grid_cpp}
\title{Build a grid from two vector (i.e. equivalent to \code{expand.grid()} in \code{R})}
\usage{
expand_grid_cpp(x, y)
}
\arguments{
\item{x}{\link{vector} first vector of numeric elements}

\item{y}{\link{vector} second vector of numeric elements}
}
\value{
\link{matrix} expanded grid of combinations
}
\description{
Build a grid from two vector (i.e. equivalent to \code{expand.grid()} in \code{R})
}
