% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_latent_cpp2}
\alias{fit_latent_cpp2}
\title{The same fit_latent_cpp, but take as argument the distance matrix directly (does not compute it by itself from coords)}
\usage{
fit_latent_cpp2(data, priors, dist, hyperpar)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{priors}{\link{list} priors: named \eqn{\mu_B},\eqn{V_r},\eqn{\Psi},\eqn{\nu}}

\item{dist}{\link{matrix} sample distance matrix}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\alpha}, second named \eqn{\phi}}
}
\value{
\link{list} posterior update parameters
}
\description{
The same fit_latent_cpp, but take as argument the distance matrix directly (does not compute it by itself from coords)
}
