% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forceSymmetry_cpp}
\alias{forceSymmetry_cpp}
\title{Function to subset data for meta-analysis}
\usage{
forceSymmetry_cpp(mat)
}
\arguments{
\item{mat}{\link{matrix} not-symmetric matrix}
}
\value{
\link{matrix} symmetric matrix (lower triangular of \code{mat} is used)
}
\description{
Function to subset data for meta-analysis
}
