% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{models_dens_latent}
\alias{models_dens_latent}
\title{Return the CV predictive density evaluations for all the model combinations}
\usage{
models_dens_latent(data, priors, coords, hyperpar, useKCV, K = 10L)
}
\arguments{
\item{data}{...}

\item{priors}{...}

\item{coords}{...}

\item{hyperpar}{...}

\item{useKCV}{...}

\item{K}{...}
}
\value{
matrix of posterior predictive density evaluations, each columns represent a different model
}
\description{
Return the CV predictive density evaluations for all the model combinations
}
