% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{models_dens_MvT}
\alias{models_dens_MvT}
\title{Return the CV predictive density evaluations for all the model combinations}
\usage{
models_dens_MvT(data, priors, coords, hyperpar, useKCV, K)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{priors}{\link{list} priors: named \eqn{\mu_B},\eqn{V_r},\eqn{\Psi},\eqn{\nu}}

\item{coords}{\link{matrix} sample coordinates for X and Y}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\alpha}, second named \eqn{\phi}}

\item{useKCV}{if \code{TRUE} K-fold cross validation is used instead of LOOCV (no \code{default})}

\item{K}{\link{integer} number of folds}
}
\value{
\link{matrix} posterior predictive density evaluations (each columns represent a different model)
}
\description{
Return the CV predictive density evaluations for all the model combinations
}
