% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PredictiveStackingWeights_cpp}
\alias{PredictiveStackingWeights_cpp}
\title{Compute the BPS stacking weights for univariate latent spatial regression model}
\usage{
PredictiveStackingWeights_cpp(
  Y,
  X,
  crd_s,
  Delta = NULL,
  Alfa = NULL,
  Fi,
  KCV = F,
  K = 10
)
}
\arguments{
\item{Y}{a "N \times 1" matrix of sample response variable}

\item{X}{a "N \times P" matrix of sample covarites}

\item{crd_s}{a "N \times 2" matrix of sample coordinates}

\item{Delta}{(univariate models) a vector of candidate values for hyperparameter \delta}

\item{Alfa}{(mulrivariate models) a vector of candidate values for hyperparameter \alpha}

\item{Fi}{a vector of candidate values for hyperparameter \phi}

\item{KCV}{a boolean flag to use K-fold cross validation instead of LOOCV (default FALSE)}

\item{K}{if KCV = TRUE, represent the number of desired K-fold}
}
\value{
a matrix with models configuration and the associated weights, and a matrix with the stacking weights
}
\description{
Compute the BPS stacking weights for univariate latent spatial regression model
}
