% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PredictiveStackingWeights_cpp2}
\alias{PredictiveStackingWeights_cpp2}
\title{Compute the BPS stacking weights for univariate latent spatial regression model}
\usage{
PredictiveStackingWeights_cpp2(Y, X, crd_s, Delta, Fi, K = 10)
}
\arguments{
\item{Y}{\link{matrix} \eqn{N \times 1} of sample response variable}

\item{X}{\link{matrix} \eqn{N \times P} matrix of sample covarites}

\item{crd_s}{\link{matrix} \eqn{N \times 2} matrix of sample coordinates}

\item{Delta}{\link{vector} (univariate models) candidate values for hyperparameter \eqn{\delta}}

\item{Fi}{\link{vector} candidate values for hyperparameter \eqn{\phi}}

\item{K}{\link{integer} if \code{KCV = TRUE}, represent the number of desired K-fold}
}
\value{
A list A list with the following components:
\item{Grid}{\link{matrix} models configuration and the associated weights}
\item{W}{\link{matrix} stacking weights}
}
\description{
Compute the BPS stacking weights for univariate latent spatial regression model
}
