% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_pred_cpp}
\alias{r_pred_cpp}
\title{Draw from the conditional posterior predictive for a set of unobserved covariates}
\usage{
r_pred_cpp(data, X_u, Rphi_s, d_u, d_us, hyperpar, poster)
}
\arguments{
\item{data}{...}

\item{X_u}{...}

\item{Rphi_s}{...}

\item{d_u}{...}

\item{d_us}{...}

\item{hyperpar}{...}

\item{poster}{...}
}
\value{
list of posterior predictive samples
}
\description{
Draw from the conditional posterior predictive for a set of unobserved covariates
}
