% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_pred_joint}
\alias{r_pred_joint}
\title{Draw from the joint posterior predictive for a set of unobserved covariates}
\usage{
r_pred_joint(data, X_u, d_u, d_us, hyperpar, poster, R)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{X_u}{\link{matrix} unobserved instances covariate matrix}

\item{d_u}{\link{matrix} unobserved instances distance matrix}

\item{d_us}{\link{matrix} cross-distance between unobserved and observed instances matrix}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\delta}, second named \eqn{\phi}}

\item{poster}{\link{list} output from \code{fit_cpp} function}
}
\value{
\link{list} posterior predictive samples
}
\description{
Draw from the joint posterior predictive for a set of unobserved covariates
}
