% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_pred_latent_cpp}
\alias{r_pred_latent_cpp}
\title{Draw from the conditional posterior predictive for a set of unobserved covariates}
\usage{
r_pred_latent_cpp(data, X_u, d_u, d_us, hyperpar, poster, beta, sigma)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{X_u}{\link{matrix} unobserved instances covariate matrix}

\item{d_u}{\link{matrix} unobserved instances distance matrix}

\item{d_us}{\link{matrix} cross-distance between unobserved and observed instances matrix}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\alpha}, second named \eqn{\phi}}

\item{poster}{\link{list} output from \code{fit_cpp} function}

\item{beta}{\link{matrix} posterior sample for \eqn{\beta}}

\item{sigma}{\link{matrix} posterior sample for \eqn{\Sigma}}
}
\value{
\link{list} posterior predictive samples
}
\description{
Draw from the conditional posterior predictive for a set of unobserved covariates
}
