% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_index}
\alias{sample_index}
\title{Function to sample integers (index)}
\usage{
sample_index(size, length, p)
}
\arguments{
\item{size}{\link{integer} dimension of the set to sample}

\item{length}{\link{integer} number of elements to sample}

\item{p}{\link{vector} sampling probabilities}
}
\value{
\link{vector} sample of integers
}
\description{
Function to sample integers (index)
}
