% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spASMK}
\alias{spASMK}
\title{Perform Accelerated Spatial Meta Kriging (ASMK)}
\usage{
spASMK(data, priors, hyperpar, K, newdata, R = 250L)
}
\arguments{
\item{data}{\link{list} three elements: first named \eqn{Y}, second named \eqn{X}, third named \eqn{crd}}

\item{priors}{\link{list} priors: named \eqn{\mu_b},\eqn{V_b},\eqn{a},\eqn{b}}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\delta}, second named \eqn{\phi}}

\item{K}{\link{integer} number of subsets}

\item{newdata}{\link{list} two elements: second named \eqn{X}, third named \eqn{crd}}

\item{R}{\link{integer} number of poterior predictive sample}
}
\value{
\link{list} posterior update parameters
}
\description{
Perform Accelerated Spatial Meta Kriging (ASMK)
}
