% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spPredict_ASMK2}
\alias{spPredict_ASMK2}
\title{Perform prediction for ASMK models - loop over prediction set}
\usage{
spPredict_ASMK2(data, X_u, priors, coords, crd_u, hyperpar, W, R, J)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{X_u}{\link{matrix} unobserved instances covariate matrix}

\item{priors}{\link{list} priors: named \eqn{\mu_b},\eqn{V_b},\eqn{a},\eqn{b}}

\item{coords}{\link{matrix} sample coordinates for X and Y}

\item{crd_u}{\link{matrix} unboserved instances coordinates}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\delta}, second named \eqn{\phi}}

\item{W}{\link{matrix} set of stacking weights}

\item{R}{\link{integer} number of desired samples}

\item{J}{\link{integer} number of desired partition of prediction set}
}
\value{
\link{list} BPS posterior predictive samples
}
\description{
Perform prediction for ASMK models - loop over prediction set
}
