% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset_data}
\alias{subset_data}
\title{Function to subset data for meta-analysis}
\usage{
subset_data(data, K)
}
\arguments{
\item{data}{\link{list} three elements: first named \eqn{Y}, second named \eqn{X}, third named \eqn{crd}}

\item{K}{\link{integer} number of desired subsets}
}
\value{
\link{list} subsets of data, and the set of indexes
}
\description{
Function to subset data for meta-analysis
}
