% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_dist}
\alias{arma_dist}
\title{Compute the Euclidean distance matrix}
\usage{
arma_dist(X)
}
\arguments{
\item{X}{\link{matrix} (tipically of \eqn{N} coordindates on \eqn{\mathbb{R}^2} )}
}
\value{
\link{matrix} distance matrix of the elements of X
}
\description{
Compute the Euclidean distance matrix
}
