% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASMK-package.R
\docType{package}
\name{ASMK-package}
\alias{ASMK}
\alias{ASMK-package}
\title{ASMK: Accelerated Spatial Meta-Kriging (ASMK)}
\description{
"Accelerated Meta-Kriging for massive Spatial dataset" by Luca Presicce and Sudipto Banerjee. This package provide the principal functions to perform accelerated meta-kriging.
}
\author{
\strong{Maintainer}: Luca Presicce \email{l.presicce@campus.unimib.it} (\href{https://orcid.org/0009-0005-7062-3523}{ORCID})

}
\keyword{internal}
