% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BPS_post_draws_latent}
\alias{BPS_post_draws_latent}
\title{Compute the BPS posterior samples given a set of stacking weights}
\usage{
BPS_post_draws_latent(data, priors, coords, hyperpar, W, R)
}
\arguments{
\item{data}{...}

\item{priors}{...}

\item{coords}{...}

\item{hyperpar}{...}

\item{W}{...}

\item{R}{...}
}
\value{
...
}
\description{
Compute the BPS posterior samples given a set of stacking weights
}
