% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_pred_cpp}
\alias{d_pred_cpp}
\title{Evaluate the density of a set of unobserved response with respect to the conditional posterior predictive}
\usage{
d_pred_cpp(data, X_u, Y_u, d_u, d_us, hyperpar, poster)
}
\arguments{
\item{data}{\link{list} two elements: first named \eqn{Y}, second named \eqn{X}}

\item{X_u}{\link{matrix} unobserved instances covariate matrix}

\item{Y_u}{\link{matrix} unobserved instances response matrix}

\item{d_u}{\link{matrix} unobserved instances distance matrix}

\item{d_us}{\link{matrix} cross-distance between unobserved and observed instances matrix}

\item{hyperpar}{\link{list} two elemets: first named \eqn{\delta}, second named \eqn{\phi}}

\item{poster}{\link{list} output from \code{fit_cpp} function}
}
\value{
\link{vector} posterior predictive density evaluations
}
\description{
Evaluate the density of a set of unobserved response with respect to the conditional posterior predictive
}
