% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dens_loocv_latent}
\alias{dens_loocv_latent}
\title{Compute the LOOCV of the density evaluations for fixed values of the hyperparameters}
\usage{
dens_loocv_latent(data, priors, coords, hyperpar)
}
\arguments{
\item{data}{\link{list} two elements: first named\eqn{Y}, second named\eqn{X}}

\item{priors}{\link{list} priors: named\eqn{\mu_B},\eqn{V_r},\eqn{\Psi},\eqn{\nu}}

\item{coords}{\link{matrix} sample coordinates for X and Y}

\item{hyperpar}{\link{list} two elemets: first named\eqn{\alpha}, second named\eqn{\phi}}
}
\value{
\link{vector} posterior predictive density evaluations
}
\description{
Compute the LOOCV of the density evaluations for fixed values of the hyperparameters
}
