% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_latent_cpp2}
\alias{fit_latent_cpp2}
\title{The same fit_latent_cpp, but take as argument the distance matrix directly (does not compute it by itself from coords)}
\usage{
fit_latent_cpp2(data, priors, dist, hyperpar)
}
\arguments{
\item{data}{two dimensional list: first named "Y", second named "X"}

\item{priors}{list with the priors: named "mu_B", "V_r", "a", "b"}

\item{hyperpar}{two dimensional list: first named "delta", second named "phi"}

\item{coords}{matrix of sample coordinates for X and Y}
}
\value{
list of posterior update parameters
}
\description{
The same fit_latent_cpp, but take as argument the distance matrix directly (does not compute it by itself from coords)
}
