% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{post_draws_latent}
\alias{post_draws_latent}
\title{Sample R draws from the posterior distributions}
\usage{
post_draws_latent(poster, R = 50L, par = FALSE, p = 2L)
}
\arguments{
\item{poster}{\link{list} output from \code{fit_cpp} function}

\item{R}{\link{integer} number of posterior samples}

\item{par}{\link{boolean} if \code{TRUE} only \eqn{\beta} and \eqn{\sigma^2} are sampled (\eqn{\omega} is omitted)}

\item{p}{\link{integer} if \code{par = TRUE}, it specifies the column number of \eqn{X}}
}
\value{
\link{list} posterior samples
}
\description{
Sample R draws from the posterior distributions
}
