% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{PredictiveStackingWeights_cpp2}
\alias{PredictiveStackingWeights_cpp2}
\title{Compute the BPS stacking weights for univariate latent spatial regression model}
\usage{
PredictiveStackingWeights_cpp2(Y, X, crd_s, Delta, Fi, K = 10)
}
\arguments{
\item{Y}{a "N \times 1" matrix of sample response variable}

\item{X}{a "N \times P" matrix of sample covarites}

\item{crd_s}{a "N \times 2" matrix of sample coordinates}

\item{Delta}{a vector of candidate values for hyperparameter \delta}

\item{Fi}{a vector of candidate values for hyperparameter \phi}

\item{K}{if KCV = TRUE, represent the number of desired K-fold}
}
\value{
a matrix with models configuration and the associated weights, and a matrix with the stacking weights
}
\description{
Compute the BPS stacking weights for univariate latent spatial regression model
}
