% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_index}
\alias{sample_index}
\title{Function to sample integers (index)}
\usage{
sample_index(size, length, p)
}
\arguments{
\item{size}{dimension of the set to sample}

\item{length}{number of elements to sample}

\item{p}{vector of probabilities}
}
\value{
vector of sampled integers
}
\description{
Function to sample integers (index)
}
